% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{NetFrac}
\alias{NetFrac}
\title{Title}
\usage{
NetFrac(x, distances = "UniFrac", paths = "single", mats = "",
  maxcores = 1, share_weight = 0)
}
\arguments{
\item{x}{The network(or tree) that will be analyzed. The network must be in the igraph format, the edges
being accessible with E(x) and vertices with V(x). The communities are under V(x)$tax and the branch weights
under E(x)$weight. These functions are applicable with igraph objects only}

\item{distances}{Distances or indices to calculate: Transfer, Transfer2, Spaths, UniFrac and Motifs. The transfer,
paths and Unifrac distances can also be calculated for trees}

\item{paths}{this parameter is used to decide whether all the shortest paths between network nodes should be
calculated ("all"), or only one of them ("single"). The last option can significantly reduce the time of
computation. For Motifs distance, decide the size of the motifs (i.e. 2 or 3)}

\item{mats}{the similarity matrix used to reconnect the network. See also \code{\link{reconnect}}}

\item{maxcores}{uses the multicore function to set up the number of cores to be used}

\item{share_weight}{weight if there are mixed communities. For now, only 0 is implemented}
}
\value{
$`Spp`\cr
A           B\cr
A 0.0000000 0.5555556\cr
B 0.5555556 0.0000000\cr

$Spep\cr
A           B\cr
A 0.0000000 0.6296296\cr
B 0.6296296 0.0000000\cr

$Spelp\cr
A           B\cr
A 0.0000000 0.6296296\cr
B 0.6296296 0.0000000\cr

$Spinp\cr
A           B\cr
A 0.0000000 0.9907407\cr
B 0.9907407 0.0000000\cr
$Transfer\cr
A           B\cr
A 0.0000000 0.6666667\cr
B 0.3333333 0.0000000\cr
}
\description{
Title
}
\examples{
NetFrac(net_a)
NetFrac(net_a,"Spaths","all")
}
