% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconnect.R
\name{reconnect}
\alias{reconnect}
\title{Connect the network using a similarity matrix}
\usage{
reconnect(graph, matrice = "")
}
\arguments{
\item{graph}{The disconnected igraph object}

\item{matrice}{The similarity matrix containing pairwise distances between the nodes}
}
\description{
Creates a fully connected network using a similarity matrix between nodes.
}
\seealso{
\code{\link{reconnect_btw}}
}
