% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconnect.R
\name{reconnect_btw}
\alias{reconnect_btw}
\title{Connect the network a graph using betweenness centrality}
\usage{
reconnect_btw(graph)
}
\arguments{
\item{graph}{The disconnected igraph object}
}
\description{
Creates a fully connected network using the betweenness centrality principle.
}
\examples{

}
\seealso{
\code{\link{reconnect}}
}
