% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_par2.r, R/dist_par3.R
\name{shortest_paths_graph}
\alias{shortest_paths_graph}
\alias{shortest_paths_graph}
\title{shortest_paths_graph}
\usage{
shortest_paths_graph(graph, v1, v2, v.mix, col1, col2, opti)

shortest_paths_graph(graph, v1, v2, v.mix, col1, col2, opti)
}
\arguments{
\item{opti}{}

\item{opti}{}
}
\description{
shortest_paths_graph

shortest_paths_graph
}
