% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSNUnifrac.R
\name{SSNUnifrac}
\alias{SSNUnifrac}
\title{Unifrac for sequence similarity network}
\usage{
SSNUnifrac(igraph, tax1 = "", tax2 = "", weight = TRUE)
}
\arguments{
  \item{igraph}{the igraph object, which is a sequence similarity network that contains at least two communities}
  \item{tax1}{community number one}
  \item{tax2}{community number two}
  \item{weight}{use the weight associated to the edges, if any. By default, this argument is set to TRUE.}
}
\description{
Calculate the UniFrac distance, based on the proportion of shared branches (here edges) over the total edges in the graph. The function uses the igraph package.
}
\examples{
net_a = load_network("exemples/network_a.txt","exemples/network_a_tax.txt")
SSNUnifrac(net_a,"A","B")
