% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_groups.R
\name{subgroup_graph}
\alias{subgroup_graph}
\title{Create a subgraph based on chosen species communities}
\usage{
subgroup_graph(graph, groups)
}
\arguments{
\item{graph}{The igraph object}

\item{groups}{The communities that should be isolated, indicated as a list of two items}
}
\description{
Creates another igraph object containing only the nodes from some chosen species
communities (two or more), and the edges connecting them.
}
\examples{
CAT_env = load_network("exemples/CAT_edges_env.csv","exemples/CAT_node_env.csv")
CAT_env1 = set_color(subgroup_graph(CAT_env, c("host", "ubiquitous")))
plot(CAT_env1)
}
