% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_dist.R
\name{Transfer}
\alias{Transfer}
\title{Algorithm for a fast calculation of the Transfer distance}
\usage{
Transfer(graph)
}
\arguments{
\item{graph}{The igraph object}
}
\description{
The trasnfer distance can be used to calculate the proportion of species/sequences of community X that have been
affected by horizontal gene transfers from species/sequences of community Y. The Transfer distance is assymetric.
If the threshold alpha (0<alpha<1) is satisfied, the distance is incremented. The operation is repeated for every
node of the network. The result is a pairwise matrix reporting the Transfer distance between each pair of species
communities.
}
\examples{
transfer2(net_a)
}
