% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_colors.r
\name{set_color}
\alias{set_color}
\title{Set colors for the communities in the network}
\usage{
set_color(graph, colors_list = "")
}
\arguments{
\item{graph}{The igraph object used to define colors. The nodes accessed with V() should have
a $tax attribute which links them to a community, so a different color can be assigned to each community.}

\item{colors_list}{Optional; the colors can be defined in a list with the community names defined with names().
Otherwise, the function will return the numbers already associated to R colors.}
}
\description{
Visualizes the network by adding different colors to different species communities, when plotting
either with R own color list or a provided color list.
}
\examples{
foo_CAT_env <- c("red","green","blue","grey")
names(foo_CAT_env) = c("host","ubiquitous","water","unknown")
CAT_env <- set_color(CAT_env, foo_CAT_env)
plot(CAT_env)
}
